+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/types/TypesTreeDumpTest.java
@@ -50,6 +50,11 @@ class TypesTreeDumpTest extends BaseTreeDumpTest {
         doTest("UnnamedPatterns");
     }
 
+    @Test
+    void testNestedLambdasAndMethodCalls() {
+        doTest("NestedLambdasAndMethodCalls");
+    }
+
     @Override
     protected @NonNull String normalize(@NonNull String str) {
         return super.normalize(str)