+++ docs/pages/release_notes.md
@@ -19,6 +19,7 @@ This is a {{ site.pmd.release_type }} release.
   * [#1860](https://github.com/pmd/pmd/issues/1860): \[ant] Reflective access warnings on java > 9 and java < 17
 * java
   * [#5293](https://github.com/pmd/pmd/issues/5293): \[java] Deadlock when executing PMD in multiple threads
+  * [#5324](https://github.com/pmd/pmd/issues/5324): \[java] Issue with type inference of nested lambdas
 
 ### 🚨 API Changes